Create Website Practice-Using CSS To Format Images
This website template is from the CSS course at www.freecodecamp.org. In this lesson I learned how to size an image using a CSS Class:
"Basic CSS: Size Your Images
CSS has a property called width that controls an element's width. Just like with fonts, we'll use px (pixels) to specify the image's width.
For example, if we wanted to create a CSS class called larger-image that gave HTML elements a width of 500 pixels, we'd use:
<style>
  .larger-image {
    width: 500px;
  }
</style>"
The challenge was:
"Create a class called smaller-image and use it to resize the image so that it's only 100 pixels wide."